team pawn-config

+++ database/migrations/2021_04_10_212640_create_lottery_prizes_table.php
@@ -16,7 +16,7 @@ class CreateLotteryPrizesTable extends Migration
         Schema::create('lottery_prizes', function (Blueprint $table) {
             $table->id();
             $table->string('name');
-            $table->double('prize_amount');
+            $table->decimal('prize_amount');
             $table->timestamps();
         });
