+++ database/migrations/2021_04_29_143701_create_sales_table.php
@@ -17,9 +17,9 @@ class CreateSalesTable extends Migration
             $table->id();
             $table->string('code')->nullable();
             $table->dateTime('dt')->nullable();
-            $table->double('gold_price_sale')->nullable();
-            $table->double('gold_price_buy')->nullable();
-            $table->double('gold_price_tax')->nullable();
+            $table->decimal('gold_price_sale')->nullable();
+            $table->decimal('gold_price_buy')->nullable();
+            $table->decimal('gold_price_tax')->nullable();
 
             $table->foreignId('customer_id')->nullable();
             $table->string('customer_name')->nullable();
@@ -27,15 +27,15 @@ class CreateSalesTable extends Migration
             $table->string('customer_tax_id')->nullable();
 
 
-            $table->double('total_price_sale')->nullable();
-            $table->double('total_price_buy')->nullable();
-            $table->double('total_amount')->nullable();
+            $table->decimal('total_price_sale')->nullable();
+            $table->decimal('total_price_buy')->nullable();
+            $table->decimal('total_amount')->nullable();
 
-            $table->double('total_wt_sale')->nullable();
-            $table->double('total_wt_buy')->nullable();
-            $table->double('total_qty_sale')->nullable();
-            $table->double('product_cost_price')->nullable();
-            $table->double('total_deposit')->nullable();
+            $table->decimal('total_wt_sale')->nullable();
+            $table->decimal('total_wt_buy')->nullable();
+            $table->decimal('total_qty_sale')->nullable();
+            $table->decimal('product_cost_price')->nullable();
+            $table->decimal('total_deposit')->nullable();
             $table->foreignId('user_id')->nullable();
             $table->string('user_name')->nullable();
             $table->foreignId('team_id');
