+++ database/migrations/2021_05_05_161509_create_gold_prices_table.php
@@ -16,16 +16,16 @@ class CreateGoldPricesTable extends Migration
         Schema::create('gold_prices', function (Blueprint $table) {
             $table->id();
             $table->dateTime('dt');
-            $table->double('gold_price_sale');
-            $table->double('gold_price_buy');
-            $table->double('gold_price_diff');
-            $table->double('gold_price_tax');
-            $table->double('spot_bid')->nullable();
-            $table->double('spot_ask')->nullable();
-            $table->double('spot_diff')->nullable();
-            $table->double('thb_bid')->nullable();
-            $table->double('thb_ask')->nullable();
-            $table->double('thb_diff')->nullable();
+            $table->decimal('gold_price_sale');
+            $table->decimal('gold_price_buy');
+            $table->decimal('gold_price_diff');
+            $table->decimal('gold_price_tax');
+            $table->decimal('spot_bid')->nullable();
+            $table->decimal('spot_ask')->nullable();
+            $table->decimal('spot_diff')->nullable();
+            $table->decimal('thb_bid')->nullable();
+            $table->decimal('thb_ask')->nullable();
+            $table->decimal('thb_diff')->nullable();
             $table->timestamps();
         });
     }
